+++ CHANGELOG.md
@@ -18,9 +18,10 @@
 * Created functions for creating standard notes and note scripts easily on the web client (#686).
 * [BREAKING] Renamed plural modules to singular (#687).
 * [BREAKING] Made `idxdb` only usable on WASM targets (#685).
-* Added fixed seed option for web client generation (#688)
+* Added fixed seed option for web client generation (#688).
 * [BREAKING] Updated `init` command in the CLI to receive a `--network` flag (#690).
 * Improved CLI error messages (#682).
+* Added account creation from component templates (#680).
 
 ### Fixes